requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# AI & OCR
openai==1.12.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# Background Tasks
celery[redis]==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.2.0
isort==5.13.2

